Restructuring, removing filepaths for the time being.
